+++ src/app/shared/enum/enumUA/tech-admin/admin-tabs.ts
@@ -4,17 +4,3 @@ export enum AdminTabsTitles {
   LawsAndRegulations = 'ENUM.NAV_BAR_NAME.LAWS_AND_REGULATIONS',
   SupportInformation = 'ENUM.NAV_BAR_NAME.SUPPORT'
 }
-
-export enum AdminTabTypes {
-  AboutPortal = 'AboutPortal',
-  MainPage = 'MainInformation',
-  LawsAndRegulations = 'LawsAndRegulations',
-  SupportInformation = 'SupportInformation'
-}
-
-export enum AdminTabsTitlesParams {
-  AboutPortal,
-  MainPage,
-  LawsAndRegulations,
-  SupportInformation
-}
