--- src/app/shared/enum/enumUA/tech-admin/create-admin.ts
@@ -1,5 +1,0 @@
-export enum CreateAdminTitle {
-  ministryAdmin = 'АДМІНА МІНІСТЕРСТВА',
-  regionAdmin = 'АДМІНА РЕГІОНА',
-  territorialCommunityAdmin = 'АДМІНА ОТГ'
-}
