+++ src/app/shell/admin-tools/data/data.component.ts
@@ -3,7 +3,7 @@ import { Component, OnInit, OnDestroy } from '@angular/core';
 import { NavigationBarService } from '../../../shared/services/navigation-bar/navigation-bar.service';
 import { AddNavPath, DeleteNavPath } from '../../../shared/store/navigation.actions';
 import { NavBarName } from '../../../shared/enum/navigation-bar';
-import { AdminTabTypes } from '../../../shared/enum/enumUA/tech-admin/admin-tabs';
+import { AdminTabTypes } from '../../../shared/enum/admins';
 
 @Component({
   selector: 'app-data',
