+++ src/app/shell/admin-tools/data/history-log/history-log-table/history-log-table.component.html
@@ -33,7 +33,7 @@
       <ng-container matColumnDef="fieldName">
         <th mat-header-cell *matHeaderCellDef mat-sort-header>{{ 'FORMS.LABELS.TYPE' | translate }}</th>
         <td mat-cell *matCellDef="let element">
-          {{ typeChange[element?.fieldName] | emptyValueTransform }}
+          {{ typeChange[element?.fieldName] | translate | emptyValueTransform }}
         </td>
       </ng-container>
