+++ src/app/shell/admin-tools/data/history-log/history-log-table/history-log-table.component.ts
@@ -1,7 +1,6 @@
 import { AfterViewInit, Component, Input, OnInit, ViewChild } from '@angular/core';
 import { MatSort } from '@angular/material/sort';
 import { MatTableDataSource } from '@angular/material/table';
-import { TypeChange } from '../../../../../shared/enum/enumUA/tech-admin/history-log';
 import { Util } from '../../../../../shared/utils/utils';
 import {
   ApplicationHistory,
@@ -12,6 +11,7 @@ import { Constants } from '../../../../../shared/constants/constants';
 import { ApplicationTitles } from '../../../../../shared/enum/enumUA/statuses';
 import { HistoryLogTypes } from '../../../../../shared/enum/history.log';
 import { Person } from '../../../../../shared/models/user.model';
+import { TypeChange } from '../../../../../shared/enum/enumUA/tech-admin/history-log';
 
 @Component({
   selector: 'app-history-log-table',
