+++ src/app/shell/admin-tools/platform/platform-info/about-info.component.ts
@@ -2,8 +2,8 @@ import { Select } from '@ngxs/store';
 import { Component } from '@angular/core';
 import { Observable } from 'rxjs';
 import { CompanyInformation } from '../../../../shared/models/сompanyInformation.model';
-import { AdminTabTypes } from '../../../../shared/enum/enumUA/tech-admin/admin-tabs';
 import { AdminState } from '../../../../shared/store/admin.state';
+import { AdminTabTypes } from '../../../../shared/enum/admins';
 @Component({
   selector: 'app-about-info',
   template: '<app-info-card [type]="aboutPortal" [platformInfo]="AboutPortal$ | async"></app-info-card>'
@@ -13,6 +13,4 @@ export class AboutInfoComponent {
 
   @Select(AdminState.AboutPortal)
   AboutPortal$: Observable<CompanyInformation>;
-
-  constructor() {}
 }
