+++ src/app/shell/admin-tools/platform/platform-info/info-card/info-card.component.ts
@@ -1,7 +1,7 @@
 import { Component, Input } from '@angular/core';
 import { Select } from '@ngxs/store';
 import { Observable } from 'rxjs';
-import { AdminTabTypes } from '../../../../../shared/enum/enumUA/tech-admin/admin-tabs';
+import { AdminTabTypes } from '../../../../../shared/enum/admins';
 import { CompanyInformation } from '../../../../../shared/models/сompanyInformation.model';
 import { AdminState } from '../../../../../shared/store/admin.state';
 
@@ -16,6 +16,4 @@ export class InfoCardComponent {
 
   @Select(AdminState.isLoading)
   isLoading$: Observable<boolean>;
-
-  constructor() {}
 }
