+++ src/app/shell/admin-tools/platform/platform-info/info-edit/info-edit.component.ts
@@ -3,7 +3,7 @@ import { Component, OnDestroy, OnInit } from '@angular/core';
 import { FormArray, FormBuilder, FormControl, FormGroup, Validators } from '@angular/forms';
 import { Select, Store } from '@ngxs/store';
 import { filter, takeUntil, tap } from 'rxjs/operators';
-import { AdminTabTypes, AdminTabsTitles } from './../../../../../shared/enum/enumUA/tech-admin/admin-tabs';
+import { AdminTabsTitles } from './../../../../../shared/enum/enumUA/tech-admin/admin-tabs';
 import { Location } from '@angular/common';
 import { Observable } from 'rxjs';
 import { ValidationConstants } from '../../../../../shared/constants/validation';
@@ -14,6 +14,7 @@ import { UpdatePlatformInfo, GetPlatformInfo } from '../../../../../shared/store
 import { AdminState } from '../../../../../shared/store/admin.state';
 import { AddNavPath } from '../../../../../shared/store/navigation.actions';
 import { CreateFormComponent } from '../../../../personal-cabinet/shared-cabinet/create-form/create-form.component';
+import { AdminTabTypes } from '../../../../../shared/enum/admins';
 
 @Component({
   selector: 'app-info-edit',
