+++ src/app/shell/admin-tools/platform/platform.component.ts
@@ -6,13 +6,12 @@ import { Subject } from 'rxjs';
 import { debounceTime, takeUntil } from 'rxjs/operators';
 import {
   AdminTabsTitles,
-  AdminTabsTitlesParams,
-  AdminTabTypes,
 } from '../../../shared/enum/enumUA/tech-admin/admin-tabs';
 import { NavBarName } from '../../../shared/enum/navigation-bar';
 import { NavigationBarService } from '../../../shared/services/navigation-bar/navigation-bar.service';
 import { GetPlatformInfo } from '../../../shared/store/admin.actions';
 import { AddNavPath, DeleteNavPath } from '../../../shared/store/navigation.actions';
+import { AdminTabsTitlesParams, AdminTabTypes } from '../../../shared/enum/admins';
 
 @Component({
   selector: 'app-platform',
