+++ src/assets/i18n/uk.json
@@ -525,7 +525,8 @@
     "NO_HISTORY": "Історії змін поки немає",
     "NO_INFO": "Інформації поки немає",
     "NO_MESSAGES": "Повідомлень поки немає",
-    "NO_NOTIFICATIONS": "У вас немає нових сповіщень"
+    "NO_NOTIFICATIONS": "У вас немає нових сповіщень",
+    "NO_SETTLEMENT": "Такого населенного пункту немає"
   },
   "NOTIFICATIONS": "Сповіщення",
   "APPLICATION_STATUSES": "Статуси заявок",
@@ -536,7 +537,8 @@
     "RESET_FILTERS": "Скинути фільтри",
     "PERIOD": "Період",
     "REMOVE_ADMIN_OPTION": "Видалення адміністратора з гуртка",
-    "ADD_ADMIN_OPTION": "Створення адміністратора"
+    "ADD_ADMIN_OPTION": "Створення адміністратора",
+    "STATUS_CHANGE_TYPE": "Зміна статусів заявок"
   },
   "STATISTICS": {
     "DATE_RANGE": "Часовий проміжок",
@@ -581,5 +583,8 @@
       "PENDING": "Отримавши статус затвердженої ліцензії, на сторінці попереднього перегляду закладу або гуртка з'явиться значок підтвердження.",
       "NOT_PROVIDED": "Отримавши статус затвердженої ліцензії, на сторінці попереднього перегляду закладу або гуртка з'явиться значок підтвердження."
     }
+  },
+  "SERVICE_MESSAGES": {
+    "NO_INFO" : "не вказано"
   }
 }
